+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Export `DecoratorType` Protocol for type-safe decorator parameter passing (#9)
+- Add `resource_predicate` parameter to `setup_resources()` for inheritance-based resource detection (#8)
 
 ## [1.0.1] - 2026-01-04